Separate running and analyzing benchmark
